refactor: move saved job CLI handlers into job module
Keep main focused on command dispatch by moving saved job list, run, and delete handlers into the shared job module alongside the saved job execution logic.

Made-with: Cursor
