+++ src/job.rs
@@ -6,7 +6,10 @@
 //! Used by both the CLI (`esdiag job run`) and the web server.
 
 use crate::{
-    data::{CollectMode, CollectSource, KnownHost, ProcessMode, SendMode, Uri, Workflow},
+    data::{
+        CollectMode, CollectSource, KnownHost, ProcessMode, SendMode, Uri, Workflow,
+        load_saved_jobs, save_saved_jobs,
+    },
     exporter::Exporter,
     processor::{
         Collector, Identifiers, Processor,
@@ -16,7 +19,98 @@ use crate::{
     uploader,
 };
 use eyre::{Result, eyre};
-use std::{path::PathBuf, sync::Arc};
+use std::{io::IsTerminal, path::PathBuf, sync::Arc};
+
+pub fn handle_job_list() -> Result<()> {
+    let jobs = load_saved_jobs()?;
+    if jobs.is_empty() {
+        return Ok(());
+    }
+
+    let hosts = KnownHost::parse_hosts_yml()?;
+
+    #[allow(clippy::literal_string_with_formatting_args)]
+    let header = format!(
+        "{:<24} {:<24} {:<16} {}",
+        "Name", "Collection target", "Processing", "Send target"
+    );
+    println!("{header}");
+    let separator: String = "-".repeat(80);
+    println!("{separator}");
+    let use_color = std::io::stdout().is_terminal();
+
+    for (name, job) in &jobs {
+        let collect_target = &job.workflow.collect.known_host;
+        let stale = !collect_target.is_empty() && !hosts.contains_key(collect_target);
+        let collect_display = if stale && use_color {
+            format!("\x1b[31m{collect_target}\x1b[0m")
+        } else {
+            collect_target.clone()
+        };
+
+        let processing = if job.workflow.process.enabled {
+            &job.workflow.process.diagnostic_type
+        } else {
+            "skipped"
+        };
+
+        let send_target = match job.workflow.send.mode {
+            SendMode::Remote => job.workflow.send.remote_target.clone(),
+            SendMode::Local => {
+                if job.workflow.send.local_target == "directory" {
+                    format!("dir:{}", job.workflow.send.local_directory)
+                } else {
+                    job.workflow.send.local_target.clone()
+                }
+            }
+        };
+
+        println!(
+            "{:<24} {:<24} {:<16} {}",
+            name, collect_display, processing, send_target
+        );
+    }
+
+    Ok(())
+}
+
+pub async fn handle_job_run(name: &str) -> Result<()> {
+    let jobs = load_saved_jobs()?;
+    let job = jobs
+        .get(name)
+        .ok_or_else(|| eyre!("Saved job '{}' not found", name))?;
+
+    let host_name = &job.workflow.collect.known_host;
+    if host_name.is_empty() {
+        return Err(eyre!(
+            "Saved job '{}' has no collection host configured",
+            name
+        ));
+    }
+
+    let host = KnownHost::get_known(host_name).ok_or_else(|| {
+        eyre!(
+            "Host '{}' referenced by job '{}' not found in hosts.yml",
+            host_name,
+            name
+        )
+    })?;
+
+    tracing::info!("Running saved job '{name}'");
+    run_saved_job(&job.workflow, job.identifiers.clone(), host).await?;
+    tracing::info!("Saved job '{name}' completed successfully");
+    Ok(())
+}
+
+pub fn handle_job_delete(name: &str) -> Result<()> {
+    let mut jobs = load_saved_jobs()?;
+    if jobs.shift_remove(name).is_none() {
+        return Err(eyre!("Saved job '{}' not found", name));
+    }
+    save_saved_jobs(&jobs)?;
+    tracing::info!("Deleted saved job '{name}'");
+    Ok(())
+}
 
 pub async fn run_saved_job(
     workflow: &Workflow,
